requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2022.12.7
click==8.1.3
Flask==2.2.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mutagen==1.46.0
pycryptodomex==3.17
PyYAML==6.0
watchdog==3.0.0
websockets==11.0
Werkzeug==2.2.3
yt-dlp==2023.3.4
